LayoutTests/media/modern-media-controls/media-controller/media-controller-fullscreen-change-expected.txt

Testing the MediaController behavior when entering and leaving fullscreen.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS internals.isChangingPresentationMode(media) became false
Media entered fullscreen
PASS mediaController.layoutTraits is an instance of MacOSLayoutTraits
PASS mediaController.layoutTraits.isFullscreen is true
PASS mediaController.controls is an instance of MacOSFullscreenMediaControls

PASS internals.isChangingPresentationMode(media) became false
Media exited fullscreen
PASS mediaController.layoutTraits is an instance of MacOSLayoutTraits
PASS mediaController.layoutTraits.isFullscreen is false
PASS mediaController.controls is an instance of MacOSInlineMediaControls

PASS successfullyParsed is true

TEST COMPLETE
